backend/requirements.txt
fastapi==0.115.11
openai-whisper
elevenlabs==1.54.0
python-dotenv==1.0.1
uuid
googletrans==4.0.2
sqlalchemy==2.0.39
python_jose==3.4.0
passlib==1.7.4
uvicorn==0.34.0
asyncio==3.4.3
python-multipart==0.0.20
bcrypt==4.3.0